requirements.txt
openai>=1.0.0
google-cloud-translate>=3.0.0 